backend/requirements.txt
alembic==1.16.5
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
cryptography==46.0.3
gunicorn==23.0.0
PyMySQL==1.1.2
